Move 'subset_dataset' to 'ops.subsetting'
